move more components to common module
