+++ common/package.json
@@ -34,6 +34,8 @@
     "react": "^17.0.2",
     "react-big-calendar": "^1.5.0",
     "react-dom": "^17.0.2",
+    "react-hook-form": "7.40",
+    "react-i18next": "11.8.5",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.5"
   },
