+++ common/src/reservation-form/RadioButtonWithImage.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import styled from "styled-components";
-import { capitalize } from "../../modules/util";
+import { capitalize } from "../../../ui/modules/util";
 
 type Props = {
   id: string;
